move retry code to its own filter
