+++ grpc.gemspec
@@ -195,6 +195,10 @@ Gem::Specification.new do |s|
   s.files += %w( src/core/ext/filters/client_channel/resolver_registry.h )
   s.files += %w( src/core/ext/filters/client_channel/resolver_result_parsing.cc )
   s.files += %w( src/core/ext/filters/client_channel/resolver_result_parsing.h )
+  s.files += %w( src/core/ext/filters/client_channel/retry_filter.cc )
+  s.files += %w( src/core/ext/filters/client_channel/retry_filter.h )
+  s.files += %w( src/core/ext/filters/client_channel/retry_service_config.cc )
+  s.files += %w( src/core/ext/filters/client_channel/retry_service_config.h )
   s.files += %w( src/core/ext/filters/client_channel/retry_throttle.cc )
   s.files += %w( src/core/ext/filters/client_channel/retry_throttle.h )
   s.files += %w( src/core/ext/filters/client_channel/server_address.cc )
